Separates out into client sentinel service
