+++ static/js/services/index.js
@@ -14,6 +14,7 @@
   require('./check-date');
   require('./child-facility');
   require('./clean-etag');
+  require('./client-sentinel');
   require('./contact-change-filter');
   require('./contact-form');
   require('./contact-save');
